LayoutTests/fast/mediastream/enumerateDevices-microphone-denied-expected.txt


PASS enumerateDevices() filters out microphones if microphone permission is prompt
PASS enumerateDevices() filters out microphones if microphone permission is denied
PASS enumerateDevices() does not filter out microphones if microphone permission is granted
